+++ tests/test_simulation_engine.py
@@ -30,9 +30,9 @@ from RUFAS.data_structures.crop_soil_to_feed_storage_connection import (
 )
 from RUFAS.input_manager import InputManager
 from RUFAS.output_manager import OutputManager
-from RUFAS.routines.field.field.field import Field
-from RUFAS.routines.field.field.manure_application import ManureApplication
-from RUFAS.routines.field.manager.field_manager import FieldManager
+from RUFAS.biophysical.field.field.field import Field
+from RUFAS.biophysical.field.field.manure_application import ManureApplication
+from RUFAS.biophysical.field.manager.field_manager import FieldManager
 from RUFAS.biophysical.manure.manure_manager import ManureManager
 from RUFAS.simulation_engine import SimulationEngine
 from RUFAS.rufas_time import RufasTime